requirements.txt
boto3==1.3.1
requests==2.8.1
elasticsearch>=5.0.0,<6.0.0
requests-aws4auth==0.9

